requirements.txt
pywebview
yt-dlp